gemini/requirements.txt
gemini-api
google-generativeai